src-yes-and-classifier/requirements.txt
absl-py==0.8.0
alabaster==0.7.12
aspy.yaml==1.3.0
astor==0.7.1
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
blis==0.2.4
boto3==1.9.205
botocore==1.12.205
bz2file==0.98
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
cfgv==2.0.1
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
cycler==0.10.0
cymem==2.0.2
decorator==4.4.0
dill==0.3.0
docutils==0.14
dopamine-rl==2.0.5
entrypoints==0.3
fairseq==0.8.0
fastBPE==0.1.0
flake8==3.7.8
flake8-bugbear==19.8.0
flake8-docstrings==1.4.0
flake8-rst-docstrings==0.0.11
Flask==1.1.1
ftfy==5.6
future==0.17.1
gast==0.3.0
gevent==1.4.0
gin-config==0.2.0
gitdb2==2.0.5
GitPython==3.0.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
googleapis-common-protos==1.6.0
greenlet==0.4.15
grpcio==1.23.0
gunicorn==19.9.0
gym==0.14.0
h5py==2.10.0
httplib2==0.13.1
identify==1.4.7
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.19
importlib-resources==1.0.2
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kfac==0.2.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mesh-tensorflow==0.0.5
mkl-fft==1.0.12
mkl-random==1.0.2
mkl-service==2.0.2
mock==3.0.5
more-itertools==7.2.0
mpmath==1.1.0
murmurhash==1.0.2
mypy==0.720
mypy-extensions==0.4.1
nltk==3.4.5
nodeenv==1.3.3
numpy==1.16.4
oauth2client==4.1.3
opencv-python==4.1.0.25
packaging==19.1
pandas==0.25.1
pexpect==4.7.0
Pillow==6.1.0
plac==0.9.6
pluggy==0.12.0
portalocker==1.5.1
pre-commit==1.18.3
preshed==2.0.1
promise==2.2.1
protobuf==3.6.1
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
py-gfm==0.1.4
py-rouge==1.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycodestyle==2.5.0
pycparser==2.19
pydocstyle==4.0.1
pyflakes==2.1.1
pyglet==1.3.2
Pygments==2.4.2
pyparsing==2.4.2
pypng==0.0.20
pytest==5.1.1
python-dateutil==2.8.0
pytorch-ignite==0.2.0
pytorch-transformers==1.0.0
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.1.0
regex==2019.6.8
requests==2.22.0
requests-mock==1.7.0
restructuredtext-lint==1.3.0
rsa==4.0
s3transfer==0.2.1
sacrebleu==1.3.7
scikit-learn==0.21.3
scipy==1.3.1
sentencepiece==0.1.82
sh==1.12.14
six==1.12.0
sklearn==0.0
smmap2==2.0.5
snowballstemmer==1.9.1
soupsieve==1.9.2
spacy==2.1.8
Sphinx==2.2.0
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
srsly==0.1.0
sympy==1.4
tb-nightly==1.15.0a20190708
tensor2tensor==1.14.0
tensorboard==1.13.1
tensorboardX==1.8
tensorflow==1.13.1
tensorflow-datasets==1.2.0
tensorflow-estimator==1.13.0
tensorflow-gan==1.0.0.dev0
tensorflow-metadata==0.14.0
tensorflow-probability==0.7.0
termcolor==1.1.0
test-tube==0.7.1
thinc==7.0.8
toml==0.10.0
torch==1.2.0
tqdm==4.36.0
typed-ast==1.4.0
typing==3.7.4.1
typing-extensions==3.7.4
uritemplate==3.0.0
urllib3==1.25.3
virtualenv==16.7.5
wasabi==0.2.2
wcwidth==0.1.7
websocket-client==0.56.0
websocket-server==0.4
Werkzeug==0.15.5
zipp==0.6.0
